docs(readme): switch documentation order
It is important to tell what this library does _before_ discussing why.
Users who navigate here will already know why they are looking for a
library, so they want to know first that this lib does what they are
looking for. Otherwise, they will not want to take the time to read the
explanation text.