refactor(validation): remove Schematize duplication
We were duplicating the Schematize type to allow for arbitrary recursion
depth. It appears that this wasn't necessary, it allowed us to get
around an issue with SchematizeItems where generating an array caused
self referential issues. By moving SchematizeItems to using an interface
with numeric keys (nearly equivalent to an array) we allow ourselves to
clean up the Schematize duplicated code.