feat: add base validation of allOf schema type
This is only a "base" implementation because it will not work for nested
schemas yet. At the moment, "allOf" must be the root level schema. This
is because typescript views the "allOf" schematizer as a circular
reference, making it necessary to manually list each depth iteration
in the base schematizer. Hopefully there will be a way around this.